Option for sorting and selecting by volume rather than chapter
Including new options to support the feature:
- '-s' to select whether to download all, last or first of the same
  numbered chapters, or to ask which to download in case of any.
- '-u' and '-g' to set preferred uploaders and groups in case of same
  numbered chapters. These take priority over the mode set with '-s'.

These options required moving requesting group names before starting to
download pages, resulting a preprocessing phase before the actual
downloading. With large amounts of chapters this could take a while so
the progress of it is also shown. The time is off the download phase so
it doesn't add much to the total process time.

Also reworked the output to show a progress bar that adapts to the
terminal width.